inserting relationship between list and music

+++ src/main/java/com/example/music/repository/PlayListRepository.java
@@ -3,5 +3,5 @@ package com.example.music.repository;
 import com.example.music.model.PlayList;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlayListRepository extends JpaRepository<PlayList, Long> {
+public interface PlayListRepository extends JpaRepository<PlayList, Integer> {
 }
